hal_402_device_mgr:  Clean up sim vs. hw mode handling
Looking forward to the day when this is launched into the real world
on its own, remove dependence on non-standard `/sim` and `/sim_mode`
parameters, and add a `sim` argument to the `hal_402_drive_mgr.launch`
file, passing the value to the node through the environment.

In the `hal_402_mgr.py` script, work these changes into
`check_for_real_hardware_setup()`, and also flatten out the code for
readability.  (Patch best viewed with git `-w` option.)