hal_402_device_mgr:  Remove self-linking code
The `hal_402_device_mgr` is primarily a HAL user component, but also
has the unusual feature of taking care of linking its own pins to
other pins.

This has already caused problems with race conditions, since it is
started in parallel with the `hal_mgr`, which is intended to be the
primary source of HAL configuration, and this component can connect
signals to other pins that `hal_mgr` also wants to connect to.  This
code had provisions to deal with that case when `hal_mgr` got there
first, but when this component got there first, `hal_mgr` HAL setup
would error out trying to connect signals to pins that were already
connected.  This race condition has become harder to paper around
since additional changes in the `hal_mgr` configuration.

The easiest fix for this is simply to make `hal_mgr` the sole source
of HAL configuration.  A single source of configuration is the
long-standing standard pattern established by HAL developers, and it
eliminates complex, fault-prone code to deal with coordinating
independent changes on the same configuration.

This patch removes the HAL configuration code from the
`hal_402_device_mgr` in preparation for it to be re-added to the
`hal_mgr` configuration.