hal_402_drive.py: fix deadlock due to unknown status word
As proposed by zultron, we can assume a race condition can
occur while reading status word pins. This can result in
an unknown status word, which will be assigned as 'unknown'

Because this 'unknown' value is only active for one cycle
of the 402 manager, in the next cycle the status word is
valid.
But due to `next_transition(self)` looking into the
transition table what the next state should be coming
from the current state the loop calling this function
can not be exited because there is never a 'True' returned
(successful transition). Hence the `elif` statement in
line 250. This will mimic a successful transition.

Thus were the target states also added in the transition
tables. (lines 71 and 83)

Also fixed: Cope with Keyerror when there is an unknown
status word. Albeit this should not happen, the 402 manager
is now hardened and can cope with this situation.