lcec module initial commit
The linuxcnc-ethercat `lcec` component for Machinekit HAL is an
EtherCAT driver that uses the IgH Master.

IgH Master exposes SDOs through an `ioctl()` interface and through a
command-line utility.  The first SDO access in this repo used the
command-line utility, and it is the same today, even though the
`ioctl()` interface might be faster and cleaner.