mgr:  Move device interface key creation to `init_devices` method
There's a chicken and egg problem in `init()`, where the parent
`init()` initializes interfaces, but `init_devices()` hasn't yet
created device-related interface keys.  This may not yet be fully
resolved, but this is a crude attempt that moves device interface key
creation to `init_devices()`.  It may still be a problem in the HAL
manager, since the HAL comp device `init()` method creates the HAL
comp, creates interfaces with HAL pins, and then marks the comp ready.