Move goal timeout from HWDeviceMgr to Device class
A device takes too long to reach its goal should trigger a "fault" to
be reported in the device, not just in the manager.  This enables
applications that watch individual devices to see the fault.  It also
lets the device report its own, more meaningful, error messages that
the manager can pass along.