cia_402:  Inherit from error device & handle `fault_desc` interface
When the CiA402 `status_word` fault bit is set, the error device class
can provide a meaningful interpretation.  Implement this by adding the
error device class as a parent class, and if there's an error code
with description, borrow that for the `fault_desc`.

Also, instead of raising an exception upon encountering a bogus status
word, set a fault and note the reason in the `fault_desc`.

Update tests after these changes.