mgr:  Improve error message handling
The top-level mgr logic was not handling errors well.  When a drive
faulted, the mgr would only note the fact that it faulted, but not the
reason.  This might have been OK, since drives report their own
faults, but it's better to have them in a single place.

Also, when the manager faulted for reasons other than an actual drive
fault, it would go into FAULT state, but then "forget" the reason.
Now the mgr retains the reason until FAULT state is exited.

Update tests for these changes.