cia_301:  Report fault if devices go offline
If devices were previously online or online+operational, report a
fault if they leave those states.  This can happen e.g. when a drive
is unplugged.

If devices were previously in fault state, it's probably because the
device went offline.  (It's not a fault for a device to start out in
offline state.)  In that case, carry the fault & fault description
feedback forward.