revert "mgr:  `async_task.py` initial commit"
This approach won't work for a number of reasons.

The `await` and `async def` syntax would have to be worked all the
way down the stack through the `device`, `config` and `command`
modules.  This would require a lot more rewriting of scripts and other
things that don't need async param operations, and would make things
cumbersome, hard to reuse, and require duplicating code.

Furthermore, while ROS2 does use an event loop, it's known that
`await` and `async def` are incompatible with `rclpy`.