cia_402:  Track status word following error bit
Add a `following_error` feedback output tied to status word
`OPERATION_MODE_SPECIFIC_2` bit 13, which represents following error
in `PP` and `CSP` modes (and homing error in `HM` mode, but that has a
separate feedback output).

The `move_error` feedback output was never connected to anything
meaningful:  in `get_feedback_pp`, it only set this output when
`move_request` active but drive not `OPERATION ENABLED`, but the
device mgr would indicate this anyway by triggering a fault.  The
output is now removed.