cia_402:  Stay in QUICK STOP ACTIVE after fault command
After a FAULT command from the mgr, an enabled drive transitions to
QUICK STOP ACTIVE, but then transitions to SWITCH ON DISABLED at the
next update.  Instead, the drive should remain in QUICK STOP ACTIVE
until a stop or start command (or drive may be set up to automatically
transition on its own).