cia_402:  Fix drive transition from quick stop
When the drive goal is `OPERATION ENABLED`, and current state is
`QUICK STOP ACTIVE`, `605Ah` "quick stop option code" determines
whether, after velocity reaches zero, the drive remains in quick stop
mode or automatically transitions to `SWITCH ON DISABLED`.  In either
case, the correct next transition is to `OPERATION ENABLED`, because
if the drive auto-transitions, the device mgr will then be looking at
where to go from `SWITCH ON DISABLED`.